requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.0.0
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
mysql-connector-python==9.1.0
numpy==2.2.6
packaging==25.0
pillow==11.0.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
pyTelegramBotAPI==4.28.0
pytest==8.4.1
requests==2.32.4
telebot==0.0.5
ttkbootstrap==1.10.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1